refactor: reorganize routers and update PC listing report functionality
