+++ src/schemas/pcListing.ts
@@ -1,7 +1,7 @@
 import { z } from 'zod'
 import { PAGINATION, CHAR_LIMITS } from '@/data/constants'
 import { HumanVerificationTokenSchema } from '@/features/human-verification/shared/schema'
-import { JsonValueSchema } from '@/schemas/common'
+import { JsonValueSchema, SortDirectionSchema } from '@/schemas/common'
 import { CreatePcListingBaseSchema } from '@/schemas/listingCreate'
 import { REVIEW_RISK_FILTERS, ReviewRiskFilterSchema } from '@/schemas/submissionRisk'
 import { ApprovalStatus, PcOs, ReportReason, ReportStatus } from '@orm'
@@ -274,6 +274,8 @@ export const UnpinPcListingCommentSchema = z.object({
 })
 
 // PC Listing Report schemas
+export const PcListingReportSortField = z.enum(['createdAt', 'updatedAt', 'status', 'reason'])
+
 export const CreatePcListingReportSchema = z.object({
   pcListingId: z.string().uuid(),
   reason: z.nativeEnum(ReportReason),
@@ -286,11 +288,17 @@ export const UpdatePcListingReportSchema = z.object({
   reviewNotes: z.string().max(1000).optional(),
 })
 
-export const GetPcListingReportsSchema = z.object({
-  status: z.nativeEnum(ReportStatus).optional(),
-  page: z.number().min(1).default(1),
-  limit: z.number().min(1).max(100).default(20),
-})
+export const GetPcListingReportsSchema = z
+  .object({
+    search: z.string().optional(),
+    status: z.nativeEnum(ReportStatus).optional(),
+    reason: z.nativeEnum(ReportReason).optional(),
+    sortField: PcListingReportSortField.optional(),
+    sortDirection: SortDirectionSchema.optional(),
+    page: z.number().min(1).default(1),
+    limit: z.number().min(1).max(100).default(20),
+  })
+  .optional()
 
 // PC Listing Verification schemas
 export const VerifyPcListingSchema = z.object({
