+++ src/server/api/routers/listings/index.ts
@@ -1,4 +1,3 @@
 export { coreRouter } from './core'
 export { commentsRouter } from './comments'
 export { adminRouter } from './admin'
-export { validateCustomFields } from './validation'
